+++ WrapSphere.cpp
@@ -15,9 +15,9 @@ void WrapSphere::compute()
     Eigen::Vector3f p = this->M * (this->point_P - this->point_O);
     Eigen::Vector3f s = this->M * (this->point_S - this->point_O);
 
-    float denom_q = p(0)*p(0) + p(1)*p(1);
-    float denom_t = s(0)*s(0) + s(1)*s(1);
-    float R = this->radius;
+    double denom_q = p(0)*p(0) + p(1)*p(1);
+    double denom_t = s(0)*s(0) + s(1)*s(1);
+    double R = this->radius;
     
     if ((denom_q - R*R < 0.0f) || (denom_t - R*R < 0.0f))
     {
@@ -25,8 +25,8 @@ void WrapSphere::compute()
         
     }
         
-    float root_q = sqrt(denom_q - R*R);
-    float root_t = sqrt(denom_t - R*R);
+    double root_q = sqrt(denom_q - R*R);
+    double root_t = sqrt(denom_t - R*R);
 
     Eigen::Vector3f q(0.0f, 0.0f, 0.0f);
     Eigen::Vector3f t(0.0f, 0.0f, 0.0f);
@@ -58,17 +58,17 @@ void WrapSphere::compute()
 
 Eigen::MatrixXf WrapSphere::getPoints(int num_points)
 {
-    float theta_q = atan(this->point_q(1) / this->point_q(0));
+    double theta_q = atan(this->point_q(1) / this->point_q(0));
     if (this->point_q(0) < 0.0f)
         theta_q += PI;
     
-    float theta_t = atan(this->point_t(1) / this->point_t(0));
+    double theta_t = atan(this->point_t(1) / this->point_t(0));
     if (this->point_t(0) < 0.0f)
         theta_t += PI;
     
     Eigen::MatrixXf points(3, num_points + 1);
     
-    float theta_s, theta_e;
+    double theta_s, theta_e;
     
     if (theta_q < theta_t)
     {
@@ -81,11 +81,11 @@ Eigen::MatrixXf WrapSphere::getPoints(int num_points)
 
     if (theta_e - theta_s > theta_s + 2*PI - theta_e)
     {
-        float tmp = theta_s; theta_s = theta_e; theta_e = tmp + 2*PI;
+        double tmp = theta_s; theta_s = theta_e; theta_e = tmp + 2*PI;
     }
 
     int col = 0;    
-    for (float i = theta_s; i <= theta_e + 0.001; 
+    for (double i = theta_s; i <= theta_e + 0.001; 
          i += (theta_e - theta_s) / num_points)
     {
         Eigen::Vector3f point = this->radius * this->M.transpose() *
